students move to graveyard when they die
